refactor: extract utillity functions into util.go
The Middleware uses utillity functions in multiple RPCs. These are
split off into a separate util.go file. This is done to have cleaner
code separation between RPCs and untilitly functions.
